feat (Cartesify-transport): add event handlers to connect method
The connect method in CartesifyTransport has been enhanced to handle specific game events such as patches, updates, initial sync, match data, and chat messages.

The following functions have been added to the connect method:
- Handle patch events
- Handle update events
- Handle initial sync events
- Handle match data events
- Handle chat events

These functions use this.cartesifyFetch to send requests to the server and process the corresponding responses. The connection status is also updated accordingly.

Additionally, the old cartesifyTransport file has been removed and replaced with a new one that follows the kebab-case naming convention.
